+++ src/Summoner/GhcVer.hs
@@ -1,10 +1,17 @@
 module Summoner.GhcVer
        ( GhcVer (..)
+       , Pvp (..)
        , showGhcVer
        , parseGhcVer
        , latestLts
        , baseVer
+       , cabalBaseVersions
        ) where
+
+import Data.List (maximum, minimum)
+
+import qualified Text.Show as Show
+
 
 -- | Represents some selected set of GHC versions.
 data GhcVer
@@ -36,11 +43,35 @@ latestLts = \case
     Ghc822  -> "11.22"
     Ghc843  -> "12.14"
 
--- | Returns base version by 'GhcVer'.
+-- | Represents PVP versioning (4 numbers).
+data Pvp = Pvp
+    { pvpFirst  :: Int
+    , pvpSecond :: Int
+    , pvpThird  :: Int
+    , pvpFourth :: Int
+    }
+
+-- | Show PVP version in a standard way: @1.2.3.4@
+instance Show Pvp where
+    show (Pvp a b c d) = intercalate "." $ map Show.show [a, b, c, d]
+
+-- | Returns base version by 'GhcVer' as 'Pvp'.
+baseVerPvp :: GhcVer -> Pvp
+baseVerPvp = \case
+    Ghc7103 -> Pvp 4 8 0 2
+    Ghc801  -> Pvp 4 9 0 0
+    Ghc802  -> Pvp 4 9 1 0
+    Ghc822  -> Pvp 4 10 1 0
+    Ghc843  -> Pvp 4 11 1 0
+
 baseVer :: GhcVer -> Text
-baseVer = \case
-    Ghc7103 -> "4.8.0.2"
-    Ghc801  -> "4.9.0.0"
-    Ghc802  -> "4.9.1.0"
-    Ghc822  -> "4.10.1.0"
-    Ghc843  -> "4.11.1.0"
+baseVer = show . baseVerPvp
+
+cabalBaseVersions :: [GhcVer] -> Text
+cabalBaseVersions []   = ""
+cabalBaseVersions [v]  = "^>= " <> baseVer v
+cabalBaseVersions ghcs = ">= " <> baseVer (minimum ghcs) <> " && < " <> upperBound
+  where
+    upperBound :: Text
+    upperBound = let Pvp{..} = baseVerPvp $ maximum ghcs in
+        show pvpFirst <> "." <> show (pvpSecond + 1)